moved close_next_date and closed_previous_date from intents to entities and make it independet from context.now by introducing a parameter for the date.
